Second Commit: Some Sprites and Class Scripts
I haven't had much time to work on this assessment up until this point, so I intend to dedicate a sizable chunk of time and effort to this project in the coming week in order to make up for time I have missed so far. I have spent time on the theoretical function of a frogger game and brainstorming using pseudocode to make sure I feel confident for the upcoming work I do.